+++ extensions/premier-league/src/hooks/index.ts
@@ -1,0 +1,15 @@
+import useFixtures from "./useFixtures";
+import useManagers from "./useManagers";
+import usePlayers from "./usePlayers";
+import useSeasons from "./useSeasons";
+import useTables from "./useTables";
+import useTeams from "./useTeams";
+
+export {
+  useFixtures,
+  useManagers,
+  usePlayers,
+  useSeasons,
+  useTables,
+  useTeams,
+};