+++ extensions/premier-league/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Premier League Changelog
 
+## [Improvements] - 2022-03-30
+
+- Added Players
+- Added Managers
+- Get seasons, clubs from API instead
+- Refactor: Use hooks, merge & remove duplicate interface
+
 ## [Initial Version] - 2022-03-24
 
 - Added Standings